managing weights and spanw count

+++ rocust_lib/src/traits.rs
@@ -18,7 +18,7 @@ pub trait HasTask {
 pub trait User: Send + Sized {
     type Shared: Shared;
 
-    fn new(_id: u16, _handler: &EventsHandler, _shared: Self::Shared) -> Self; // TODO: pass test config and a test controller to be able to stop the test based on some user defined conditions
+    fn new(_id: u64, _handler: &EventsHandler, _shared: Self::Shared) -> Self; // TODO: pass test config and a test controller to be able to stop the test based on some user defined conditions
     fn on_start(&mut self, _handler: &EventsHandler) {}
     fn on_stop(&mut self, _handler: &EventsHandler) {}
 }
